Move html charts to folder
